Introduzione test genere

+++ controllers/gestione/Rimuovi genere.php
@@ -1,11 +1,20 @@
 <?php
 include_once "../../php/core.php";
-allowOnlyGestore();
-$genere_id = @$_GET["genere_id"];
+//allowOnlyGestore();
+
+$utente = Auth::getLoggedUser();
+if (is_null($utente) or !$utente->isGestore()) {
+    Testing::redirect("/");
+    return;
+}
+$genere_id = trim(@$_POST["genere_id"]);
 $ff = new FormFeedbacker();
 $genere_dao = GenereDAOFactory::getGenereDAO();
-if ($genere_dao->delete($genere_id))
-	header("Location: /");
+if ($genere_dao->delete($genere_id)) {
+    Testing::redirect("/");
+    return;
+}
 else
-	$ff->bug();
+    $ff->bug();
+
 $ff->process();
